feat: move file existence check functions to utils and clean up stage 5
